fix cellsize distribution
