+++ src/utils/plotSpecs/generateCellSizeDistributionHistogram.js
@@ -1,0 +1,192 @@
+const generateSpec = (config, plotData) => {
+  let legend = null;
+  const minHigh = 8800;
+  const minUnknown = 10800;
+  const generateStatus = `(datum.u < ${minUnknown}) ? 'low' : (datum.u >${minHigh}) ? 'high' : 'unknown'`;
+  const coloringExpressionPlot = "(datum.bin1 < 8800) ? 'low' : (datum.bin1 >10800) ? 'high' : 'unknown'";
+
+  legend = !config.legend.enabled ? {} : [
+    {
+      fill: 'color',
+      orient: config.legend.position,
+      title: 'Quality',
+      labelFont: config.fontStyle.font,
+      titleFont: config.fontStyle.font,
+      encode: {
+        title: {
+          update: {
+            fontSize: { value: 14 },
+          },
+        },
+        labels: {
+          interactive: true,
+          update: {
+            fontSize: { value: 12 },
+            fill: { value: 'black' },
+          },
+        },
+        symbols: {
+          update: {
+            stroke: { value: 'transparent' },
+          },
+        },
+        legend: {
+          update: {
+            stroke: { value: '#ccc' },
+            strokeWidth: { value: 1.5 },
+          },
+        },
+      },
+    }];
+
+  return {
+    $schema: 'https://vega.github.io/schema/vega/v5.json',
+    width: config.dimensions.width,
+    height: config.dimensions.height,
+    autosize: { type: 'fit', resize: true },
+
+    padding: 5,
+
+    data: [
+      {
+        name: 'plotData',
+        values: plotData,
+        transform: [
+          {
+            type: 'formula',
+            as: 'status',
+            expr: generateStatus,
+          },
+        ],
+      },
+      {
+        name: 'binned',
+        source: 'plotData',
+        transform: [
+          {
+            type: 'bin',
+            field: 'u',
+            extent: [0, 17000],
+            step: config.binStep,
+            nice: false,
+          },
+          {
+            type: 'aggregate',
+            key: 'bin0',
+            groupby: ['bin0', 'bin1'],
+            fields: ['bin0'],
+            ops: ['count'],
+            as: ['count'],
+          },
+          {
+            type: 'formula',
+            as: 'color',
+            expr: coloringExpressionPlot,
+          },
+        ],
+      },
+    ],
+
+    scales: [
+      {
+        name: 'xscale',
+        type: 'linear',
+        range: 'width',
+        domain: [1000, 17000],
+      },
+      {
+        name: 'yscale',
+        type: 'linear',
+        range: 'height',
+        round: true,
+        domain: { data: 'binned', field: 'count' },
+        zero: true,
+        nice: true,
+      },
+      {
+        name: 'color',
+        type: 'ordinal',
+        range: ['green', '#f57b42', 'grey'],
+        domain: {
+          data: 'binned',
+          field: 'color',
+          sort: true,
+        },
+      },
+    ],
+    axes: [
+      {
+        orient: 'bottom',
+        scale: 'xscale',
+        grid: true,
+        zindex: 1,
+        title: { value: config.axes.xAxisText },
+        titleFont: { value: config.fontStyle.font },
+        labelFont: { value: config.fontStyle.font },
+        titleFontSize: { value: config.axes.titleFontSize },
+        labelFontSize: { value: config.axes.labelFontSize },
+        offset: { value: config.axes.offset },
+        gridOpacity: { value: config.axes.gridOpacity / 20 },
+      },
+      {
+        orient: 'left',
+        scale: 'yscale',
+        grid: true,
+        zindex: 1,
+        title: { value: config.axes.yAxisText },
+        titleFont: { value: config.fontStyle.font },
+        labelFont: { value: config.fontStyle.font },
+        titleFontSize: { value: config.axes.titleFontSize },
+        labelFontSize: { value: config.axes.labelFontSize },
+        offset: { value: config.axes.offset },
+        gridOpacity: { value: config.axes.gridOpacity / 20 },
+      },
+    ],
+    marks: [
+      {
+        type: 'rect',
+        from: { data: 'binned' },
+        encode: {
+          enter: {
+            x: { scale: 'xscale', field: 'bin0' },
+            x2: {
+              scale: 'xscale',
+              field: 'bin1',
+            },
+            y: { scale: 'yscale', field: 'count' },
+            y2: { scale: 'yscale', value: 0 },
+            stroke: { value: 'black' },
+            strokeWidth: { value: 0.5 },
+            fill: {
+              scale: 'color',
+              field: 'color',
+            },
+          },
+        },
+      },
+      {
+        type: 'rule',
+        encode: {
+          update: {
+            x: { scale: 'xscale', value: config.minCellSize },
+            y: { value: 0 },
+            y2: { field: { group: 'height' } },
+            strokeWidth: { value: 2 },
+            strokeDash: { value: [8, 4] },
+            stroke: { value: 'red' },
+          },
+        },
+      },
+    ],
+    legends: legend,
+    title: {
+      text: { value: config.title.text },
+      anchor: { value: config.title.anchor },
+      font: { value: config.fontStyle.font },
+      dx: 10,
+      fontSize: { value: config.title.fontSize },
+    },
+  };
+};
+
+export default generateSpec;
