+++ src/utils/plotSpecs/generateCellSizeDistributionKneePlot.js
@@ -1,0 +1,170 @@
+const generateSpec = (config, plotData) => {
+  let legend = null;
+  const minHigh = 2500;
+  const minUnknown = 2300;
+  const generateStatus = `(datum.u < ${minUnknown}) ? 'low' : (datum.u >${minHigh}) ? 'high' : 'unknown'`;
+
+  legend = !config.legend.enabled ? null : [
+    {
+      fill: 'color',
+      orient: config.legend.position,
+      title: 'Quality',
+      labelFont: config.fontStyle.font,
+      titleFont: config.fontStyle.font,
+      encode: {
+        title: {
+          update: {
+            fontSize: { value: 14 },
+          },
+        },
+        labels: {
+          interactive: true,
+          update: {
+            fontSize: { value: 12 },
+            fill: { value: 'black' },
+          },
+        },
+        symbols: {
+          update: {
+            stroke: { value: 'transparent' },
+          },
+        },
+        legend: {
+          update: {
+            stroke: { value: '#ccc' },
+            strokeWidth: { value: 1.5 },
+          },
+        },
+      },
+    }];
+
+  return {
+    $schema: 'https://vega.github.io/schema/vega/v5.json',
+    width: config.dimensions.width,
+    height: config.dimensions.height,
+    autosize: { type: 'fit', resize: true },
+    padding: 5,
+
+    data: [
+      {
+        name: 'plotData',
+        values: plotData,
+        transform: [
+          {
+            type: 'formula',
+            as: 'status',
+            expr: generateStatus,
+          },
+          {
+            type: 'filter',
+            expr: 'datum.u > 0 && datum.rank > 0',
+          },
+          {
+            type: 'formula',
+            as: 'logRankValue',
+            expr: '(log(datum.rank) / LN10)*2200',
+          },
+        ],
+      },
+    ],
+
+    scales: [
+      {
+        name: 'xscale',
+        type: 'log',
+        range: 'width',
+        domain: { data: 'plotData', field: 'u' },
+      },
+      {
+        name: 'yscale',
+        type: 'linear',
+        range: 'height',
+        nice: true,
+        domain: { data: 'plotData', field: 'logRankValue' },
+      },
+      {
+        name: 'color',
+        type: 'ordinal',
+        range: ['green', '#f57b42', 'grey'],
+        domain: {
+          data: 'plotData',
+          field: 'status',
+        },
+      },
+    ],
+
+    axes: [
+      {
+        orient: 'bottom',
+        scale: 'xscale',
+        tickCount: 5,
+        grid: true,
+        zindex: 1,
+        title: { value: config.axes.xAxisText },
+        titleFont: { value: config.fontStyle.font },
+        labelFont: { value: config.fontStyle.font },
+        titleFontSize: { value: config.axes.titleFontSize },
+        labelFontSize: { value: config.axes.labelFontSize },
+        offset: { value: config.axes.offset },
+        gridOpacity: { value: config.axes.gridOpacity / 20 },
+      },
+      {
+        orient: 'left',
+        scale: 'yscale',
+        grid: true,
+        zindex: 1,
+        title: { value: config.axes.yAxisText },
+        titleFont: { value: config.fontStyle.font },
+        labelFont: { value: config.fontStyle.font },
+        titleFontSize: { value: config.axes.titleFontSize },
+        labelFontSize: { value: config.axes.labelFontSize },
+        offset: { value: config.axes.offset },
+        gridOpacity: { value: config.axes.gridOpacity / 20 },
+      },
+    ],
+
+    marks: [
+      {
+        type: 'area',
+        from: { data: 'plotData' },
+        encode: {
+          enter: {
+            x: { scale: 'xscale', field: 'u' },
+            y: { scale: 'yscale', field: 'logRankValue' },
+            y2: { scale: 'yscale', value: 0 },
+            fill: {
+              scale: 'color',
+              field: 'status',
+            },
+          },
+          update: {
+            fillOpacity: { value: 1 },
+          },
+        },
+      },
+      {
+        type: 'rule',
+        encode: {
+          update: {
+            x: { scale: 'xscale', value: config.minCellSize },
+            y: { value: 0 },
+            y2: { field: { group: 'height' } },
+            strokeWidth: { value: 2 },
+            strokeDash: { value: [8, 4] },
+            stroke: { value: 'red' },
+          },
+        },
+      },
+    ],
+    legends: legend,
+    title: {
+      text: { value: config.title.text },
+      anchor: { value: config.title.anchor },
+      font: { value: config.fontStyle.font },
+      dx: { value: config.title.dx },
+      fontSize: { value: config.title.fontSize },
+    },
+  };
+};
+
+export default generateSpec;
